requirements.txt
pyperclip==1.9.0
streamlit-extras==0.4.3
openpyxl==3.1.5


